+++ CHANGES.md
@@ -4,6 +4,7 @@
     * `Twin::representer_class.representable_attrs` is now `Twin::definitions`.
 * `Disposable::Twin::Schema` is now `Disposable::Rescheme`. Renamed its `:representer_from` option to `:definitions_from`.
 * `twin: Twin::Song` is the only way to specify an external twin. To reduce complexity, I've removed the lambda version of `:twin`.
+* Added `:exclude_properties` to `Rescheme`.
 
 # 0.1.14